Refactor cli code
