--- src/cli/mod.rs
@@ -1,2 +1,0 @@
-pub mod args;
-pub mod config;
