+++ src/lib.rs
@@ -1,6 +1,6 @@
 pub mod cli;
 
-use cli::config::Config;
+use cli::Config;
 
 use std::panic;
 use std::path::Path;
@@ -148,7 +148,7 @@ mod tests {
         let text = "Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est laborum.";
         let file = "qr_tmp.png";
 
-        let config = cli::config::Config {
+        let config = cli::Config {
             input: Some(text),
             output: Some(file),
             read: false,
@@ -170,7 +170,7 @@ mod tests {
             thread_rng().sample_iter(&Alphanumeric).take(30).collect();
         let file = "qr_tmp_random.png";
 
-        let config = cli::config::Config {
+        let config = cli::Config {
             input: Some(&text),
             output: Some(file),
             read: false,
